+++ src/main/java/ru/biryukov/dto/ProductDTO.java
@@ -16,14 +16,14 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class ProductDTO {
 
-    private int id;
+    private Long id;
 
     @NotBlank(message = "Name cannot be empty")
     @Size(max = 255, message = "Product name must be no more than 255 characters")
     private String name;
 
     @Size(max = 4096, message = "Product description must be no more than 4096 characters")
-    private String description= "";
+    private String description = "";
 
     @PositiveOrZero(message = "Product price cannot be negative")
     @Digits(integer = 10, fraction = 2, message = "Product price must have up to 2 digits after the decimal point")
